fix the cli install script for windows

+++ fission/install-windows-cli.ps1
@@ -1,3 +1,5 @@
 go install
-rm $Env:GOPATH\bin\fission-test.exe
+if (Test-Path $Env:GOPATH\bin\fission-test.exe -PathType Leaf) {
+    rm $Env:GOPATH\bin\fission-test.exe
+}
 mv $Env:GOPATH\bin\fission.exe $Env:GOPATH\bin\fission-test.exe
